+++ History.md
@@ -4,6 +4,7 @@
 
 - WARNING: if you are using a PHP 7.3 or older, then update to PHP 8.0 or newer first before installing this update!
 - In speaking lists, it is now possible to remove and reorder speakers
+- Votings can now be reordered
 
 ### Version 4.10.1 (2022-07-01)